pyproject.toml
[tool.poetry]
name = "kubectl-forwarding"
version = "0.1.0"
description = ""
authors = ["Maxim Osipov <kingco6ra@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
pylama = "^8.4.1"
mypy = "^1.8.0"
mashumaro = "^3.12"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
isort = "^5.13.2"
flake8 = "^7.0.0"
toml-sort = "^0.23.1"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pre-commit = "*"
toml-sort = "*"
vulture = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
